\name{matchFeat-package}
\alias{matchFeat-package}

\docType{package}
\title{
\packageTitle{matchFeat}
}
\description{
\packageDescription{matchFeat}
}
\details{


}

\author{
Author: \packageAuthor{matchFeat}
Maintainer: \packageMaintainer{matchFeat}
}

\references{
Degras (2021) "Scalable Feature Matching Across Large Data Collections". 
\url{https://arxiv.org/abs/2101.02035} \cr
Wright (2015). Coordinate descent algorithms.
\url{https://arxiv.org/abs/1502.04759} \cr
McLachlan and Krishnan (2008). \emph{The EM Algorithm and Extensions}
}
