\name{matchFeats-package}
\alias{matchFeats-package}

\docType{package}
\title{
\packageTitle{matchFeats}
}
\description{
\packageDescription{matchFeats}
}
\details{


}

\author{
Author: \packageAuthor{matchFeats}
Maintainer: \packageMaintainer{matchFeats}
}

\references{
Degras and Chen (2020). One-to-one feature matching under association ambiguity. \cr
\url{https://www.crcv.ucf.edu/projects/GMMCP-Tracker} \cr
Wright (2015). Coordinate descent algorithms.
\url{https://arxiv.org/abs/1502.04759} \cr
McLachlan and Krishnan (2008). \emph{The EM Algorithm and Extensions}
}
