\name{matchFeat-package}
\alias{matchFeat-package}

\docType{package}
\title{
\packageTitle{matchFeat}
}
\description{
\packageDescription{matchFeat}
}
\details{


}

\author{
Author: \packageAuthor{matchFeat}
Maintainer: \packageMaintainer{matchFeat}
}

\references{
Degras (2020). One-to-one feature matching under association ambiguity. \cr
\url{https://www.crcv.ucf.edu/projects/GMMCP-Tracker} \cr
Wright (2015). Coordinate descent algorithms.
\url{https://arxiv.org/abs/1502.04759} \cr
McLachlan and Krishnan (2008). \emph{The EM Algorithm and Extensions}
}
